Branch Update

+++ src/Login.jsx
@@ -55,9 +55,13 @@ const Login = () => {
                     break;
                 case 200:
                     const rep = await res.data
-                    localStorage.setItem('id', rep._id)
-                    localStorage.setItem('token', rep.mainToken)
-                    navigate('/')
+                    if (rep.confirmed) {
+                        localStorage.setItem('id', rep._id)
+                        localStorage.setItem('token', rep.mainToken)
+                        navigate('/')
+                    } else {
+                        navigate("/verify")
+                    }
                     break;
                 default:
                     break;
@@ -127,7 +131,7 @@ const Login = () => {
     if (reg === "sent") {
         return (
             <div className="loginMainDiv">
-                <div style={{ position: "relative" }} className="loginDiv">
+                <div style={{ position: "relative", height: "600px" }} className="loginDiv">
                     <div className="lockImage">
                         <img src={mail} alt="" />
                     </div>
